refactor(input,table): remove dead rich-clipboard store, fix forward-declared singleton
- W3: delete the orphaned InternalClipboard store (setRichClipboard /
  consumeRichClipboard / clearRichClipboard had no production callers). The
  RichBlockData / RichSegment type imports in ClipboardHandler and
  PasteRichBlockHandler now come from model/RichBlockData directly. Removes the
  module's misleading unit test and the moot "stale data" PasteHandler test.
- W11: move TableNavigation's activeContextMenu module singleton above its first
  use instead of relying on it being declared further down the file.
